added opting to byzantine generals example; one now decides to do one of: - "attack" :: |packets| == |links| && attack packets > retreat packets - "retreat" :: |packets| == |links| && attack packets < retreat packets - "unknown" :: |packets| == |links| && attack packets == retreat packets - "not yet known" :: |packets| < |links|
this can be seen in the variables panel

+++ trunk/src/vidis/modules/byzantineGenerals/APacket.java
@@ -9,13 +9,13 @@ import vidis.data.annotation.DisplayColor;
 public abstract class APacket extends AUserPacket {
 	private static Logger logger = Logger.getLogger(APacket.class);
 	
-	private double id;
+	private int id;
 	
 	public APacket() {
-		id = Math.random() * Double.MAX_VALUE;
+		id = (int)(Math.random() * Double.MAX_VALUE);
 	}
 	
-	public APacket(double id) {
+	public APacket(int id) {
 		this.id = id;
 	}
 	
@@ -37,7 +37,7 @@ public abstract class APacket extends AUserPacket {
 		return getPacketType().getColor();
 	}
 	
-	public double getId() {
+	public int getId() {
 		return id;
 	}
